Only copy file if it exists.

+++ setup.py
@@ -633,7 +633,7 @@ class my_build_ext(build_ext):
             if name1 is not None:
                 src = os.path.join(self.build_temp, os.path.dirname(ext.sources[0]), name1)
                 dst = os.path.join(self.build_temp, name2)
-            if os.path.abspath(src) != os.path.abspath(dst):
+            if os.path.abspath(src) != os.path.abspath(dst) and os.path.isfile(src):
                 self.copy_file(src, dst)
 
         for ext in W32_exe_files:
